+++ library/src/main/java/com/sarang/torang/usecase/comments/AddCommentLikeUseCase.kt
@@ -4,5 +4,5 @@ package com.sarang.torang.usecase.comments
  * Request comment like/delete like
  */
 interface AddCommentLikeUseCase {
-    suspend fun invoke(commentId: Int): Int
+    suspend fun invoke(commentId: Long): Int
 }
