+++ library/src/main/java/com/sarang/torang/usecase/comments/DeleteCommentUseCase.kt
@@ -1,5 +1,5 @@
 package com.sarang.torang.usecase.comments
 
 interface DeleteCommentUseCase {
-    suspend fun delete(commentId: Int)
+    suspend fun delete(commentId: Long)
 }
